Move localstorage methods to own module
