+++ src/project.js
@@ -1,21 +1,22 @@
+import { updateProjectLS } from "./localstorage"
+
 // prototype for project
 
 const projectProto = {
   addTask(task) {
-    this.tasks.push(task)
+    this.tasks.push(task);
   },
   deleteTask(taskTitle) {
-    const taskIndex = this.tasks.findIndex( task => task.title === taskTitle);
+    const taskIndex = this.tasks.findIndex((task) => task.title === taskTitle);
     return this.tasks.splice(taskIndex, 1);
   },
   getTasks() {
     return this.tasks;
   },
 
-  displayTasks() {   
-  
+  displayTasks() {
     const projectTasks = this.tasks;
-    
+
     const container = document.createElement("div");
     container.className = "toDos-div";
 
@@ -27,16 +28,25 @@ const projectProto = {
 
     projectTasks.forEach((task) => {
       let taskLink = document.createElement("li");
+      let taskDelete = document.createElement("button");
+      let trashIcon = document.createElement("i");
+      trashIcon.className = "fas fa-trash";
       taskLink.className = `task-li-${task.id}`;
       taskLink.innerHTML = task.title;
+      taskDelete.appendChild(trashIcon);
+      taskLink.onclick = () => {
+        this.deleteTask(task.title)
+        // this.displayTasks();
+      }
+      taskLink.appendChild(taskDelete);
       ulContainer.appendChild(taskLink);
-    }); 
+    });
     container.appendChild(projectTitle);
     container.appendChild(ulContainer);
-    
-    return  container;
-}
-}
+
+    return container;
+  },
+};
 
 // factory function for project
 
@@ -47,4 +57,4 @@ const projectFactory = (title, tasks = []) => {
   return obj;
 };
 
-export { projectFactory, projectProto }
+export { projectFactory, projectProto };
